+++ packages/athena_mysql/CHANGELOG.md
@@ -4,3 +4,6 @@
 
 ## 1.0.2
 - bump athena_sql version
+
+## 1.0.3
+- add `mapColumnOrTable` to driver