+++ packages/athena_postgres/CHANGELOG.md
@@ -6,3 +6,6 @@
 - fix transaction types
 - fix connection map variables
 - fix column extension
+
+## 1.0.3
+- add `mapColumnOrTable` to driver